Add index params check in proxy
